data_generator/requirements.txt
fire
kafka-python
numpy